+++ CHANGES.rst
@@ -10,6 +10,7 @@ CHANGELOG
 - separate ignore filters from mount option filters
 - allow to match multiple attributes against a device (AND-wise)
 - allow to overwrite udiskies default handleability settings
+- raise exception if --config file doesn't exist
 
 
 0.7.0